pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "VDS"
version = "1.1"
authors = [
    {name = "Trieu Viet Hung", email = "hungtrieu0701@gmail.com"},
]
description = "VDS"
requires-python = "==3.8.*"

dependencies = [
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp38-cp38-linux_x86_64.whl ; platform_system=='Linux'",
    "torch @ https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp38-cp38-win_amd64.whl ; platform_system=='Windows'",
    "torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.9.1%2Bcpu-cp38-cp38-linux_x86_64.whl ; platform_system=='Linux'",
    "torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.9.1%2Bcpu-cp38-cp38-win_amd64.whl ; platform_system=='Windows'",
    "lap05==0.5.1",
    "numpy==1.23.0",
    "altgraph==0.17.3",
    "certifi==2023.5.7",
    "charset-normalizer==3.1.0",
    "click==8.1.4",
    "colorama==0.4.6",
    "Cython==0.29.35",
    "cython_bbox@git+https://github.com/samson-wang/cython_bbox",
    "cytoolz==0.12.1",
    "debugpy==1.6.7.post1",
    "dnspython==2.3.0",
    "et-xmlfile==1.1.0",
    "filelock==3.12.2",
    "idna==3.4",
    "Jinja2==3.1.2",
    "MarkupSafe==2.1.3",
    "mpmath==1.3.0",
    "networkx==3.1",
    "onnxruntime==1.15.1",
    "opencv-python==4.7.0.72",
    "openpyxl==3.1.2",
    "packaging==23.1",
    "pandas==2.0.2",
    "pefile==2023.2.7",
    "Pillow==9.5.0",
    "pymongo==4.3.3",
    "PyQt5==5.15.9",
    "PyQt5-sip==12.12.1",
    "PyQtChart==5.15.6",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "PyYAML==6.0",
    "requests==2.31.0",
    "scipy==1.10.1",
    "six==1.16.0",
    "sympy==1.12",
    "toolz==0.12.0",
    "typing_extensions==4.6.3",
    "tzdata==2023.3",
    "urllib3==2.0.3",
    "xarray==0.21.1",
    "bcrypt==4.0.1"
]

[tool.setuptools.packages.find]
where=["core"]
